+++ application/config/autoload.php
@@ -58,7 +58,7 @@ $autoload['packages'] = array();
 |
 |	$autoload['libraries'] = array('user_agent' => 'ua');
 */
-$autoload['libraries'] = array('parser');
+$autoload['libraries'] = array('parser', 'session');
 
 /*
 | -------------------------------------------------------------------
@@ -132,4 +132,5 @@ $autoload['language'] = array();
 |
 |	$autoload['model'] = array('first_model' => 'first');
 */
-$autoload['model'] = array('planes', 'fleets', 'airlines', 'flights');
+
+$autoload['model'] = array('planesEntity', 'planeEntity', 'fleetEntity', 'airportEntity', 'airlineEntity', 'flightEntity', 'regionEntity', 'planes', 'fleets', 'airports', 'airlines', 'flights', 'regions');