+++ application/config/config.php
@@ -6,6 +6,7 @@ $config['menu_choices'] = array(
         array('name' => 'Homepage', 'link' => '/'),
         array('name' => 'Fleet', 'link' => '/fleet'),
         array('name' => 'Flights', 'link' => '/flight'),
+        array('name' => 'Bookings', 'link' => '/booking'),
     )
 );
 
@@ -378,7 +379,7 @@ $config['encryption_key'] = '';
 $config['sess_driver'] = 'files';
 $config['sess_cookie_name'] = 'ci_session';
 $config['sess_expiration'] = 7200;
-$config['sess_save_path'] = NULL;
+$config['sess_save_path'] = '/tmp'; 
 $config['sess_match_ip'] = FALSE;
 $config['sess_time_to_update'] = 300;
 $config['sess_regenerate_destroy'] = FALSE;