+++ application/models/Planes.php
@@ -1,6 +1,7 @@
 <?php
+
     /*
-        Model for plane.
+        Model for planes.
         Gets all plane data from json from server
     */
     class Planes extends CI_Model {
@@ -10,26 +11,36 @@
         // This hereby is a constructor
         public function __construct () {
             parent::__construct ();
-            $ch = curl_init();
-            curl_setopt($ch, CURLOPT_SSL_VERIFYPEER, false);
-            curl_setopt($ch, CURLOPT_RETURNTRANSFER, true);
-            curl_setopt($ch, CURLOPT_URL, 'http://wacky.jlparry.com/info/airplanes');
-            $result = curl_exec($ch);
-            curl_close($ch);
-            $this -> data = json_decode($result, true);
+            $this -> data = WackyAPI::getAirplanes ();
+            $this -> convert_to_planes_array ();
+        }
+
+        // Convert data recieved in array format from server to an array containing plane objects
+        private function convert_to_planes_array () {
+            $records = array();
+            foreach ($this -> data as $key => $record)
+                array_push ($records, PlanesEntity::create_plane_from_arr ($record));
+            $this -> data = $records;
         }
 
         // Returns all planes from server
-        public function all() 
-        {
+        public function all() {
             return $this -> data;
         }
 
         // Returns a specific airplane
         public function get($which) {
             foreach ($this -> data as $key => $record)
-                if ($record["id"] == $which)
+                if ($record -> id == $which)
                     return $record;
+            return null;
+        }
+
+        // Returns a specific airplane
+        public function get_with_view($which) {
+            foreach ($this -> data as $key => $record)
+                if ($record -> id == $which)
+                    return $record -> getViewArray();
             return null;
         }